Finally have the DeltaFan behavior somewhat properly, but PID controller is not working.

+++ firmware/lib/task/DeltaFans.h
@@ -88,7 +88,7 @@ public:
   void printRPMS();
   void motorControl(int s);
 
-  void PWMMotorControl(int s, int m);
+  void PWMMotorControl(float s, int m);
   void update(float pwm_ratio);
 };
